# api/e2etests/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: contact@mightycode.fr
    url: https://mightycode.fr/
  description: |
    This is a prototype of CPOO Project's front/back API.
  title: CPOO Server API
  version: 0.0.1
servers:
- url: http://localhost:8080/
security:
- CookieAuth: []
...
paths:
  /user/signup:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
        required: true
      responses:
        "200":
          description: Success
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
          description: Error
      security: []
      tags:
      - authentication
      x-content-type: application/json
      x-accepts: application/json
  /user/signin:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
        required: true
      responses:
        "200":
          description: Success
          headers:
            Set-Cookie:
              explode: false
              schema:
                type: string
              style: simple
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
          description: Error
      security: []
      tags:
      - authentication
      x-content-type: application/json
      x-accepts: application/json
  /user/signout:
    post:
      responses:
        "200":
          description: Success
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
          description: Error
      tags:
      - authentication
      x-accepts: application/json
  /user/{username}:
    delete:
      description: Delete the account of the user matching the given username. Only
        the administrator can use this endpoint.
      parameters:
      - description: username of the user account to delete.
        explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Success
        "404":
          description: "Not found: User account not found."
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
          description: Error
      summary: Delete a user account
      tags:
      - administration
      x-accepts: application/json
...
  /user/currentuser:
    post:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
          description: Success
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
          description: Error
      tags:
      - authentication
      x-accepts: application/json
...
  /user/changepwd:
    patch:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordDTO'
        required: true
      responses:
        "200":
          description: Success
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
          description: Error
      tags:
      - user settings
      x-content-type: application/json
      x-accepts: application/json
...
components:
  schemas:
    UserDTO:
      example:
        password: pass
        email: alice@mymail.com
        username: alice
      properties:
        username:
          example: alice
          type: string
        email:
          example: alice@mymail.com
          type: string
        password:
          example: pass
          type: string
      required:
      - email
      - password
      - username
      type: object
...
    ChangePasswordDTO:
      example:
        oldpassword: oldpassword
        newpassword: newpassword
      properties:
        oldpassword:
          example: oldpassword
          type: string
        newpassword:
          example: newpassword
          type: string
      type: object
...
    ErrorDTO:
      description: Error information
      properties:
        timestamp:
          description: Timestamp of the error
          example: 2023-10-03T14:53:59.870+00:00
          type: string
        status:
          description: HTTP response status
          example: 401
          type: integer
        error:
          description: HTTP response error
          example: Unauthorized
          type: string
        message:
          description: HTTP response message
          example: Bad credentials
          type: string
        path:
          description: Path of the endpoint called
          example: /user/signin
          type: string
      type: object
  securitySchemes:
    CookieAuth:
      description: Use a session cookie to authenticate (see /login).
      in: cookie
      name: JSESSIONID
      type: apiKey
...